--- TinkoffChat/CoreData/Channel/Channel+CoreDataClass.swift
@@ -1,15 +1,0 @@
-//
-//  Channel+CoreDataClass.swift
-//  
-//
-//  Created by Vera on 05.04.2020.
-//
-//
-
-import Foundation
-import CoreData
-
-@objc(Channel)
-public class Channel: NSManagedObject {
-
-}
