+++ TinkoffChat/CoreLayer/Modules/Channel/Channel+CoreDataClass.swift
