+++ TinkoffChat/CoreLayer/Modules/Channel/Channel+CoreDataProperties.swift
