+++ TinkoffChat/CoreLayer/Modules/Message/Message+CoreDataProperties.swift
