+++ TinkoffChat/CoreLayer/Modules/User/User+CoreDataProperties.swift
