+++ TinkoffChat/PresentationLayer/Modules/Extensions/UIColor.swift
@@ -1,0 +1,30 @@
+//
+//  UIColor.swift
+//  TinkoffChat
+//
+//  Created by Vera on 12.04.2020.
+//  Copyright © 2020 Vera. All rights reserved.
+//
+
+import UIKit
+
+extension UIColor{
+    //softpink
+    static let mainColor = UIColor(red:0.38, green:0.49, blue:0.55, alpha:1.0)//UIColor(red:0.67, green:0.62, blue:0.64, alpha:1.0)
+    static let mainLightColor = UIColor(red:0.38, green:0.49, blue:0.55, alpha:0.2)
+    //static let mainLightColor = UIColor(red:0.90, green:0.85, blue:0.85, alpha:1.0)
+    
+    static let yellowLight = UIColor(red:1.00, green:1.00, blue:0.55, alpha:1.0)
+    static let deepBlue = UIColor(red:0.10, green:0.33, blue:0.48, alpha:0.5)
+    
+    //#d299c2→#fef9d7 wild apple
+    static let pink = UIColor(red:0.82, green:0.60, blue:0.76, alpha:1.0)
+    static let greyYellow = UIColor(red:1.00, green:0.98, blue:0.84, alpha:1.0)
+    
+    //Turkish palet
+    static let unmellowYellow = UIColor(red:1.00, green:0.98, blue:0.40, alpha:1.0)
+    static let balticSea = UIColor(red:0.24, green:0.24, blue:0.24, alpha:1.0)
+    static let blueGrey500 = UIColor(red:0.38, green:0.49, blue:0.55, alpha:1.0)
+    static let blueGrey800 = UIColor(red:0.22, green:0.28, blue:0.31, alpha:1.0)
+    static let blueGrey900 = UIColor(red:0.15, green:0.20, blue:0.22, alpha:1.0)
+}
