+++ TinkoffChat/PresentationLayer/Modules/Extensions/UIFont.swift
@@ -1,0 +1,23 @@
+//
+//  UIFont.swift
+//  TinkoffChat
+//
+//  Created by Vera on 12.04.2020.
+//  Copyright © 2020 Vera. All rights reserved.
+//
+
+import UIKit
+
+extension UIFont{
+
+    class Play {
+        let Bold = UIFont(name: "Play-Bold", size: UIFont.labelFontSize)
+        let Regular = UIFont(name: "Play-Regular", size: UIFont.labelFontSize)
+    }
+    
+    class RobotoSlab{
+        let ExtraLight = UIFont(name: "RobotoSlab-ExtraLight", size: UIFont.labelFontSize)
+        let Light = UIFont(name: "RobotoSlab-Light", size: UIFont.labelFontSize)
+        let Thin = UIFont(name: "RobotoSlab-Thin", size: UIFont.labelFontSize)
+    }
+}
