+++ TinkoffChat/PresentationLayer/Modules/Extensions/UIStackView.swift
@@ -1,0 +1,18 @@
+//
+//  UIStackView.swift
+//  TinkoffChat
+//
+//  Created by Vera on 12.04.2020.
+//  Copyright © 2020 Vera. All rights reserved.
+//
+
+import UIKit
+
+extension UIStackView {
+    func addBackground(color: UIColor) {
+        let subView = UIView(frame: bounds)
+        subView.backgroundColor = color
+        subView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        insertSubview(subView, at: 0)
+    }
+}
