+++ TinkoffChat/PresentationLayer/Modules/Extensions/UIViewController.swift
@@ -1,0 +1,21 @@
+//
+//  UIViewController.swift
+//  TinkoffChat
+//
+//  Created by Vera on 12.04.2020.
+//  Copyright © 2020 Vera. All rights reserved.
+//
+
+import UIKit
+
+extension UIViewController {
+    func hideKeyboardWhenTappedAround() {
+        let tap: UITapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(UIViewController.dismissKeyboard))
+        tap.cancelsTouchesInView = false
+        view.addGestureRecognizer(tap)
+    }
+
+    @objc func dismissKeyboard() {
+        view.endEditing(true)
+    }
+}
