+++ TinkoffChat/PresentationLayer/Modules/General/ChannelListViewController/ConversationCell.swift
