+++ TinkoffChat/ServicesLayer/Assembly/ServicesAssembly.swift
@@ -1,0 +1,9 @@
+//
+//  ServicesAssembly.swift
+//  TinkoffChat
+//
+//  Created by Vera on 12.04.2020.
+//  Copyright © 2020 Vera. All rights reserved.
+//
+
+import Foundation
