+++ TinkoffChat/ServicesLayer/ConversationCellModel.swift
@@ -8,15 +8,22 @@
 
 import Foundation
 
+struct ChannelModel {
+    let identifier: String
+    let name: String?
+    let lastMessage: String?
+    let lastActivity: Date?
+}
+
 struct ConversationCellModel {
-    let channel: ChannelOld
+    let channel: ChannelModel
     let hasUnreadMessage : Bool
 }
 
 protocol ConversationCellConfiguration {
     associatedtype ConversationCellModel
     
-    var channel : ChannelOld {get set}
+    var channel : ChannelModel {get set}
     var hasUnreadMessage : Bool?{get set}
     
     func configure(with model: ConversationCellModel)
