+++ TinkoffChat/Supporting Files/Assets.xcassets/AppIcon.appiconset/Contents.json
