+++ TinkoffChat/Supporting Files/Assets.xcassets/AppIcon.appiconset/Icon-60@2x.png
