+++ TinkoffChat/Supporting Files/Assets.xcassets/AppIcon.appiconset/Icon-60@3x.png
