+++ TinkoffChat/Supporting Files/Assets.xcassets/AppIcon.appiconset/Icon1024.ong.png
