+++ TinkoffChat/Supporting Files/Assets.xcassets/AppIcon.appiconset/Icon40.png
