+++ TinkoffChat/Supporting Files/Assets.xcassets/AppIcon.appiconset/Icon60-1.png
