+++ TinkoffChat/Supporting Files/Assets.xcassets/AppIcon.appiconset/Icon60.png
