+++ TinkoffChat/Supporting Files/Assets.xcassets/AppIcon.appiconset/Icon80_.png
