+++ TinkoffChat/Supporting Files/Assets.xcassets/AppIcon.appiconset/Icon87.png
