+++ TinkoffChat/Supporting Files/Assets.xcassets/Contents.json
