+++ TinkoffChat/Supporting Files/Info.plist
