+++ TinkoffChat/Supporting Files/TinkoffChat.xcdatamodeld/TinkoffChat.xcdatamodel/contents
